+++ src/main/java/com/otilm/core/util/CertificateRequestUtils.java
@@ -1,10 +1,7 @@
 package com.otilm.core.util;
 
 import com.otilm.api.exception.CertificateRequestException;
-import com.otilm.api.model.client.attribute.RequestAttribute;
-import com.otilm.api.model.common.attribute.v2.content.StringAttributeContentV2;
 import com.otilm.api.model.core.enums.CertificateRequestFormat;
-import com.otilm.core.attribute.CsrAttributes;
 import com.otilm.core.model.request.CertificateRequest;
 import com.otilm.core.model.request.CrmfCertificateRequest;
 import com.otilm.core.model.request.Pkcs10CertificateRequest;
@@ -15,14 +12,12 @@ import org.bouncycastle.util.io.pem.PemObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.security.auth.x500.X500Principal;
 import java.io.IOException;
 import java.io.StringWriter;
 import java.security.*;
 import java.security.spec.InvalidKeySpecException;
 import java.security.spec.X509EncodedKeySpec;
 import java.util.Base64;
-import java.util.List;
 
 public class CertificateRequestUtils {
     private static final Logger logger = LoggerFactory.getLogger(CertificateRequestUtils.class);
@@ -71,87 +66,6 @@ public class CertificateRequestUtils {
             }
         }
         return publicKeyObject;
-    }
-
-    public static X500Principal buildSubject(List<RequestAttribute> attributes) {
-
-        // Get the data for the attributes
-        String commonName = AttributeDefinitionUtils.getSingleItemAttributeContentValue(
-                        CsrAttributes.COMMON_NAME_ATTRIBUTE_NAME,
-                        attributes,
-                        StringAttributeContentV2.class)
-                .getData();
-
-        String organizationalUnit = AttributeDefinitionUtils.getSingleItemAttributeContentValue(
-                        CsrAttributes.ORGANIZATION_UNIT_ATTRIBUTE_NAME,
-                        attributes,
-                        StringAttributeContentV2.class)
-                .getData();
-
-        String organization = AttributeDefinitionUtils.getSingleItemAttributeContentValue(
-                        CsrAttributes.ORGANIZATION_ATTRIBUTE_NAME,
-                        attributes,
-                        StringAttributeContentV2.class)
-                .getData();
-
-        String locality = AttributeDefinitionUtils.getSingleItemAttributeContentValue(
-                        CsrAttributes.LOCALITY_ATTRIBUTE_NAME,
-                        attributes,
-                        StringAttributeContentV2.class)
-                .getData();
-
-        String state = AttributeDefinitionUtils.getSingleItemAttributeContentValue(
-                        CsrAttributes.STATE_ATTRIBUTE_NAME,
-                        attributes,
-                        StringAttributeContentV2.class)
-                .getData();
-
-        String country = AttributeDefinitionUtils.getSingleItemAttributeContentValue(
-                        CsrAttributes.COUNTRY_ATTRIBUTE_NAME,
-                        attributes,
-                        StringAttributeContentV2.class)
-                .getData();
-
-        StringBuilder nameBuilder = new StringBuilder();
-        if (commonName != null) {
-            nameBuilder.append("CN=").append(escapeSpecialCharacters(commonName));
-        }
-        if (organizationalUnit != null) {
-            nameBuilder.append(", OU=").append(escapeSpecialCharacters(organizationalUnit));
-        }
-        if (organization != null) {
-            nameBuilder.append(", O=").append(escapeSpecialCharacters(organization));
-        }
-        if (locality != null) {
-            nameBuilder.append(", L=").append(escapeSpecialCharacters(locality));
-        }
-        if (state != null) {
-            nameBuilder.append(", ST=").append(escapeSpecialCharacters(state));
-        }
-        if (country != null) {
-            nameBuilder.append(", C=").append(escapeSpecialCharacters(country));
-        }
-        return new X500Principal(nameBuilder.toString());
-    }
-
-
-    // Escape special characters according to RFC 2253
-    private static String escapeSpecialCharacters(String inputString) {
-        // Escape one of the characters ",", "+", """, "\", "<", ">" or ";"
-        final String[] specialCharacters = {"\\", ",", "+", "\"", "<", ">" , ";"};
-        for (String specialCharacter : specialCharacters) {
-            if (inputString.contains(specialCharacter)) {
-                inputString = inputString.replace(specialCharacter, "\\" + specialCharacter);
-            }
-        }
-
-        // Escape a space character occurring at the end of the string
-        if (inputString.endsWith(" ")) inputString = inputString.substring(0, inputString.length() -1 ) + "\\ ";
-
-        // Escape a space or "#" character occurring at the beginning of the string
-        if (inputString.startsWith("#") || inputString.startsWith(" ")) return "\\" + inputString;
-
-        return inputString;
     }
 
     /**
